Update README.md
The current project is to create a website with links to a few different recipes. I will use all of the HTML skills I have acquired and learn how to use them in a real world experience. By the end of this project I will further my knowledge of HTML and be able to create a HTML project and push it to GitHub.